+++ database/migrations/2020_08_14_023546_create_rel_trabajador_escolaridad.php
@@ -16,10 +16,10 @@ class CreateRelTrabajadorEscolaridad extends Migration
         Schema::create('rel_trabajador_escolaridad', function (Blueprint $table) {
             $table->Increments('id')->unsigned();
             $table->smallInteger('trabajador_id')->unsigned();
-            $table->smallInteger('grado_academico_id')->unsigned()->nullable();
-            $table->smallInteger('nombre_estudio_id')->unsigned()->nullable();
+            $table->mediumInteger('grado_academico_id')->unsigned()->nullable();
+            $table->mediumInteger('nombre_estudio_id')->unsigned()->nullable();
             $table->string('otro_nombre_estudio', 254)->nullable();
-            $table->smallInteger('institucion_id')->unsigned()->nullable();
+            $table->mediumInteger('institucion_id')->unsigned()->nullable();
             $table->string('otro_nombre_institucion', 254)->nullable();
             $table->smallInteger('cedula')->default(0)->unsigned()->comments("1 = si, 0 = no");
             $table->smallInteger('no_cedula')->unsigned()->nullable();
